requirements.txt
flask
transformers
sentence-transformers
faiss-cpu
pymupdf